+++ .eslintignore
@@ -1,1 +1,1 @@
-app/scripts/lib/extension.js
+app/scripts/lib/extension-instance.js
